+++ veecle-os-examples/common/src/actors/ping_pong.rs
@@ -3,7 +3,7 @@
 use core::fmt::Debug;
 use futures::future::FutureExt;
 use serde::{Deserialize, Serialize};
-use veecle_os::runtime::{InitializedReader, Never, Reader, Storable, Writer};
+use veecle_os::runtime::{Never, Reader, Storable, Writer};
 use veecle_os::telemetry::{error, info};
 
 #[derive(Debug, Clone, PartialEq, Eq, Default, Storable, Deserialize, Serialize)]
@@ -19,20 +19,19 @@ pub struct Pong {
 /// An actor that writes `Ping { i }` and waits for `Pong`.
 /// Additionally, it validates that `Pong { value == i + 1 }` for `i = 0..`.
 #[veecle_os::runtime::actor]
-pub async fn ping_actor(mut ping: Writer<'_, Ping>, pong: Reader<'_, Pong>) -> Never {
+pub async fn ping_actor(mut ping: Writer<'_, Ping>, mut pong: Reader<'_, Pong>) -> Never {
     let mut value = 0;
     info!("[PING TASK] Sending initial", ping = i64::from(value));
     ping.write(Ping { value }).await;
     value += 1;
 
-    info!("[PING TASK] Waiting for pong");
-    let mut pong = pong.wait_init().await;
     loop {
         info!("[PING TASK] Waiting for pong");
-        pong.wait_for_update().await.read(|pong| {
+        pong.read_updated(|pong| {
             info!("[PING TASK] Pong received", pong = i64::from(pong.value));
             assert_eq!(pong.value, value);
-        });
+        })
+        .await;
         info!("[PING TASK] Sending", ping = i64::from(value));
         ping.write(Ping { value }).await;
         value += 1;
@@ -41,17 +40,16 @@ pub async fn ping_actor(mut ping: Writer<'_, Ping>, pong: Reader<'_, Pong>) -> N
 
 /// An actor that reads `Ping`, replies with `Pong { ping + 1 }` and waits for the next `Ping`.
 #[veecle_os::runtime::actor]
-pub async fn pong_actor(
-    mut pong: Writer<'_, Pong>,
-    mut ping: InitializedReader<'_, Ping>,
-) -> Never {
+pub async fn pong_actor(mut pong: Writer<'_, Pong>, mut ping: Reader<'_, Ping>) -> Never {
     loop {
         info!("[PONG TASK] Waiting for ping");
 
-        let value = ping.wait_for_update().await.read(|ping| {
-            info!("[PONG TASK] Ping received", ping = i64::from(ping.value));
-            ping.value + 1
-        });
+        let value = ping
+            .read_updated(|ping| {
+                info!("[PONG TASK] Ping received", ping = i64::from(ping.value));
+                ping.value + 1
+            })
+            .await;
 
         let data = Pong { value };
         info!("[PONG TASK] Sending", pong = i64::from(data.value));