+++ veecle-os-runtime-macros/tests/ui/actor/static_reference.rs
@@ -5,7 +5,7 @@ fn verify_static(_: &'static u8) {}
 
 #[veecle_os_runtime_macros::actor]
 async fn macro_test_actor(
-    reader: veecle_os_runtime::Reader<'_, Reference<'static>>,
+    mut reader: veecle_os_runtime::Reader<'_, Reference<'static>>,
     _writer: veecle_os_runtime::Writer<'_, Reference<'static>>,
 ) -> veecle_os_runtime::Never {
     verify_static(reader.read_cloned().unwrap().0);