+++ veecle-os-runtime/src/lib.rs
@@ -28,28 +28,26 @@
 //! }
 //!
 //! #[veecle_os_runtime::actor]
-//! async fn ping_actor(mut ping: Writer<'_, Ping>, pong: Reader<'_, Pong>) -> Never {
+//! async fn ping_actor(mut ping: Writer<'_, Ping>, mut pong: Reader<'_, Pong>) -> Never {
 //!     let mut value = 0;
 //!     ping.write(Ping { value }).await;
 //!
-//!     let mut pong = pong.wait_init().await;
 //!     loop {
 //!         ping.write(Ping { value }).await;
 //!         value += 1;
 //!
-//!         pong.wait_for_update().await.read(|pong| {
+//!         pong.read_updated(|pong| {
 //!             println!("Pong: {}", pong.value);
-//!         });
+//!         }).await;
 //! #       // Exit the application to allow doc-tests to complete.
 //! #       std::process::exit(0);
 //!     }
 //! }
 //!
 //! #[veecle_os_runtime::actor]
-//! async fn pong_actor(mut pong: Writer<'_, Pong>, ping: Reader<'_, Ping>) -> Never {
-//!     let mut ping = ping.wait_init().await;
+//! async fn pong_actor(mut pong: Writer<'_, Pong>, mut ping: Reader<'_, Ping>) -> Never {
 //!     loop {
-//!         let ping = ping.wait_for_update().await.read_cloned();
+//!         let ping = ping.read_updated_cloned().await;
 //!         println!("Ping: {}", ping.value);
 //!
 //!         let data = Pong { value: ping.value };
@@ -110,7 +108,7 @@ pub mod memory_pool;
 
 pub use self::actor::{Actor, StoreRequest, actor};
 pub use self::datastore::{
-    CombinableReader, CombineReaders, ExclusiveReader, InitializedReader, Reader, Storable, Writer,
+    CombinableReader, CombineReaders, ExclusiveReader, Reader, Storable, Writer,
 };
 
 /// Internal exports for proc-macro and `macro_rules!` purposes.