+++ veecle-os-runtime/tests/sensor-actuator.rs
@@ -5,9 +5,8 @@ use core::sync::atomic::{AtomicUsize, Ordering};
 use std::future::poll_fn;
 use std::sync::Mutex;
 use std::task::Poll;
-use veecle_os_runtime::Never;
 
-use veecle_os_runtime::{CombineReaders, InitializedReader, Storable, Writer};
+use veecle_os_runtime::{CombineReaders, Never, Reader, Storable, Writer};
 
 static SENSOR_VALIDATION_COUNT: AtomicUsize = AtomicUsize::new(0);
 static PRINTER_OUTPUT: Mutex<String> = Mutex::new(String::new());
@@ -36,59 +35,57 @@ async fn sensor_actor(mut sensor_writer: Writer<'_, Sensor>) -> Never {
 
 #[veecle_os_runtime::actor]
 async fn sensor_validation(
-    mut sensor_reader: InitializedReader<'_, Sensor>,
-    mut actuator_data_reader: InitializedReader<'_, ActuatorData>,
+    mut sensor_reader: Reader<'_, Sensor>,
+    mut actuator_data_reader: Reader<'_, ActuatorData>,
 ) -> Never {
     for expected_sensor in (0..).cycle() {
-        let combined_reader = (&mut sensor_reader, &mut actuator_data_reader);
+        let mut combined_reader = (&mut sensor_reader, &mut actuator_data_reader);
         combined_reader.read(|(a, b)| {
             println!("{:?}, {:?}", a, b);
         });
-        sensor_reader.wait_for_update().await.read(|&sensor| {
-            assert_eq!(sensor, expected_sensor);
-            SENSOR_VALIDATION_COUNT.fetch_add(1, Ordering::SeqCst);
-        })
+        sensor_reader
+            .read_updated(|&sensor| {
+                assert_eq!(sensor, expected_sensor);
+                SENSOR_VALIDATION_COUNT.fetch_add(1, Ordering::SeqCst);
+            })
+            .await
     }
     unreachable!("The endless loop should never end.");
 }
 
 #[veecle_os_runtime::actor]
 async fn actuator(
-    mut sensor_reader: InitializedReader<'_, Sensor>,
+    mut sensor_reader: Reader<'_, Sensor>,
     mut actuator_data_writer: Writer<'_, ActuatorData>,
 ) -> Never {
     loop {
-        let value = sensor_reader
-            .wait_for_update()
-            .await
-            .read(|sensor| sensor + 10);
+        let value = sensor_reader.read_updated(|sensor| sensor + 10).await;
         actuator_data_writer.write(value).await
     }
 }
 
 #[veecle_os_runtime::actor]
-async fn actuator_validation(
-    mut actuator_data_reader: InitializedReader<'_, ActuatorData>,
-) -> Never {
+async fn actuator_validation(mut actuator_data_reader: Reader<'_, ActuatorData>) -> Never {
     for expected_actuator in (10..).cycle() {
         actuator_data_reader
-            .wait_for_update()
-            .await
-            .read(|&actuator| {
+            .read_updated(|&actuator| {
                 assert_eq!(actuator, expected_actuator);
-            });
+            })
+            .await;
     }
     unreachable!("The endless loop should never end.");
 }
 
 #[veecle_os_runtime::actor]
-async fn printer(mut actuator_data_reader: InitializedReader<'_, ActuatorData>) -> Never {
+async fn printer(mut actuator_data_reader: Reader<'_, ActuatorData>) -> Never {
     use std::fmt::Write;
 
     loop {
-        actuator_data_reader.wait_for_update().await.read(|value| {
-            writeln!(&mut PRINTER_OUTPUT.lock().unwrap(), "{value:?}").unwrap();
-        });
+        actuator_data_reader
+            .read_updated(|value| {
+                writeln!(&mut PRINTER_OUTPUT.lock().unwrap(), "{value:?}").unwrap();
+            })
+            .await;
     }
 }